rename file makeFIeldsRequest to fields
