+++ public/jira-oidc-helpers/fields.ts
